feat(methods): move reference loading to separate method
References were loaded in the 'read' and 'select' methods. This is now moved to a separate 'expandReferences' method. This method can be used for external storage engines. Also, the logic when and what to load is improved. Before this change, a full read was executed for every referenced item, even if only the id was requested. And it was always loading all references within the reference, even if this data was not requested at all. The 'expandReferences' method only loads the requested data, and skips the read if only the id was requested. This can be a significant performance improvement when a lot of references are used.