feat(scripts): evaluate expressions according to BIDMAS rules
BREAKING CHANGE: expressions were always evaluated greedy. The right operand consumes as much of the expression as possible, leading to 2 * 3 + 4 resulting in 14, where 10 was expected. This behevior is now fixed. Although unlikely, it may break existing scripts relying on the greedy behavior. There is no change for scripts using brackets to work around the greedy parsing.