feat(methods): return null for 'and' and 'or' when left operand is false
BREAKING CHANGE: the old behavior was to return the right operand when the left operand is true for 'and' and false for 'or'. The new behavior is to return null in such cases. This behavior does not affect conditional logic, but is more convenient when composing objects. For example, '/name and uppercase /name' will now evaluate to null when name is empty, which indicates a missing value. False is not appropriate here, because name is not a boolean.